requirements.txt
# HTTP client
httpx==0.28.1

# Testing
pytest==8.3.5
pytest-asyncio==0.25.3

# Data validation
pydantic==2.10.6
pydantic-settings==2.8.1

# Structured logging
structlog==25.1.0

# Terminal rendering
rich==13.9.4

# Data aggregation
pandas==2.2.3

# LLM-as-judge
anthropic==0.49.0

# Judge prompt templates
jinja2==3.1.5

# Environment config
python-dotenv==1.0.1

# PII detection
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.8.4
